+++ apps/frontend/src/pages/Admin/DriverApprovals.tsx
@@ -1,0 +1,5 @@
+const DriverApprovals = () => {
+  return <div>Driver Approvals</div>;
+};
+
+export default DriverApprovals;
