+++ apps/frontend/src/pages/Public/About.tsx
