+++ apps/frontend/src/pages/Public/FooterPages/Safety.tsx
