+++ apps/frontend/src/pages/Public/Privacy.tsx
