+++ apps/frontend/src/pages/Public/Terms.tsx
