--- apps/frontend/src/pages/User/Assessment.tsx
@@ -1,9 +1,0 @@
-const Assessment = () => {
-  return (
-    <div>
-      English Assessment page
-    </div>
-  );
-};
-
-export default Assessment;
