+++ apps/frontend/src/pages/User/Bookings.tsx
@@ -1,0 +1,5 @@
+const Bookings = () => {
+  return <div>Bookings page</div>;
+};
+
+export default Bookings;
