+++ apps/frontend/src/router/adminSideRoutes.tsx
@@ -1,16 +1,32 @@
+import Analytics from "@/pages/Admin/Analytics";
 
 export const adminSideRoutes = [
-    {
-      title: "Dashboard",
-      items: [
-        {
-          title: "Analytics",
-          url: "/admin/analytics"
-        },
-        {
-          title: "Project Structure",
-          url: "#",
-        },
-      ],
-    },    
-]
+  {
+    title: "Dashboard",
+    items: [
+      {
+        title: "Analytics",
+        url: "/admin/analytics",
+        component: Analytics,
+      },
+      {
+        title: "Project Structure",
+        url: "#",
+      },
+    ],
+  },
+  {
+    title: "Dashboard",
+    items: [
+      {
+        title: "Bookings",
+        url: "/user/bookings",
+        component: Analytics,
+      },
+      {
+        title: "Project Structure",
+        url: "#",
+      },
+    ],
+  },
+];
